Implemented Changes Discussed with Reviewers
