+++ src/main/java/org/apache/cassandra/sidecar/MainModule.java
@@ -18,10 +18,7 @@
 
 package org.apache.cassandra.sidecar;
 
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.io.InputStream;
-import java.nio.charset.StandardCharsets;
 import java.util.Collections;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
@@ -32,7 +29,6 @@ import org.slf4j.LoggerFactory;
 import com.google.inject.AbstractModule;
 import com.google.inject.Provides;
 import com.google.inject.Singleton;
-import com.google.inject.name.Named;
 import io.netty.handler.codec.http.HttpResponseStatus;
 import io.vertx.core.Vertx;
 import io.vertx.core.VertxOptions;
@@ -50,7 +46,6 @@ import org.apache.cassandra.sidecar.cassandra40.Cassandra40Factory;
 import org.apache.cassandra.sidecar.cluster.InstancesConfig;
 import org.apache.cassandra.sidecar.common.ApiEndpointsV1;
 import org.apache.cassandra.sidecar.common.CassandraVersionProvider;
-import org.apache.cassandra.sidecar.common.NodeSettings;
 import org.apache.cassandra.sidecar.common.dns.DnsResolver;
 import org.apache.cassandra.sidecar.common.utils.ValidationConfiguration;
 import org.apache.cassandra.sidecar.logging.SidecarLoggerHandler;
@@ -70,6 +65,7 @@ import org.apache.cassandra.sidecar.routes.sstableuploads.SSTableImportHandler;
 import org.apache.cassandra.sidecar.routes.sstableuploads.SSTableUploadHandler;
 import org.apache.cassandra.sidecar.utils.ChecksumVerifier;
 import org.apache.cassandra.sidecar.utils.MD5ChecksumVerifier;
+import org.apache.cassandra.sidecar.utils.SidecarVersionProvider;
 import org.apache.cassandra.sidecar.utils.TimeProvider;
 import org.jboss.resteasy.plugins.server.vertx.VertxRegistry;
 import org.jboss.resteasy.plugins.server.vertx.VertxRequestHandler;
@@ -249,10 +245,10 @@ public class MainModule extends AbstractModule
     @Provides
     @Singleton
     public Configuration configuration(CassandraVersionProvider cassandraVersionProvider,
-                                       @Named("SidecarVersion") String sidecarVersion) throws IOException
+                                       SidecarVersionProvider sidecarVersionProvider) throws IOException
     {
         final String confPath = System.getProperty("sidecar.config", "file://./conf/config.yaml");
-        return YAMLSidecarConfiguration.of(confPath, cassandraVersionProvider, sidecarVersion);
+        return YAMLSidecarConfiguration.of(confPath, cassandraVersionProvider, sidecarVersionProvider.sidecarVersion());
     }
 
     @Provides
@@ -272,10 +268,10 @@ public class MainModule extends AbstractModule
     @Provides
     @Singleton
     public CassandraVersionProvider cassandraVersionProvider(DnsResolver dnsResolver,
-                                                             @Named("SidecarVersion") String sidecarVersion)
+                                                             SidecarVersionProvider sidecarVersionProvider)
     {
         CassandraVersionProvider.Builder builder = new CassandraVersionProvider.Builder();
-        builder.add(new Cassandra40Factory(dnsResolver, sidecarVersion));
+        builder.add(new Cassandra40Factory(dnsResolver, sidecarVersionProvider.sidecarVersion()));
         return builder.build();
     }
 
@@ -323,25 +319,8 @@ public class MainModule extends AbstractModule
 
     @Provides
     @Singleton
-    @Named("SidecarVersion")
-    public String sidecarVersion()
+    public SidecarVersionProvider sidecarVersionProvider()
     {
-        String resource = "/sidecar.version";
-        try (InputStream input = NodeSettings.class.getResourceAsStream(resource);
-             ByteArrayOutputStream output = new ByteArrayOutputStream())
-        {
-            byte[] buffer = new byte[32];
-            int length;
-            while ((length = input.read(buffer)) >= 0)
-            {
-                output.write(buffer, 0, length);
-            }
-            return output.toString(StandardCharsets.UTF_8.name());
-        }
-        catch (Exception exception)
-        {
-            LOGGER.error("Failed to retrieve Sidecar version from resource {}", resource, exception);
-        }
-        return "unknown";
+        return new SidecarVersionProvider("/sidecar.version");
     }
 }
